prevent circular dependency by moving analytics event
